Created BookHandle Module.
