+++ book_handle.rb
@@ -1,0 +1,28 @@
+require './book.rb'
+
+module BookHandle
+   # List all books
+   def list_books
+    puts '*' * 40
+    if @books.empty?
+      puts 'The book list is empty. Create new book!'
+    else
+      @books.each_with_index do |book, index|
+        puts "#{index}. Title: #{book.title} Authored by: #{book.author}"
+      end
+    end
+  end
+
+  # Create a book
+  def create_book
+    puts
+    print 'Title: '
+    title = gets.chomp.strip.capitalize
+    puts
+    print 'Author: '
+    author = gets.chomp.strip.capitalize
+    @books.push(Book.new(title, author))
+    puts "Title: #{title} - Author: #{author}. Created successfully!"
+  end
+
+end
